Split costs frontend in new/old parts
