+++ frontend/app/openproject-costs-app.js
@@ -26,63 +26,5 @@
 // See doc/COPYRIGHT.rdoc for more details.
 //++
 
-var CostsByTypeDisplayField = require('./components/wp-display/field-types/wp-display-costs-by-type-field.module').CostsByTypeDisplayField;
-var CurrencyDisplayField = require('./components/wp-display/field-types/wp-display-currency-field.module').CurrencyDisplayField;
-
-// Register Budget as select inline edit
-angular
-  .module('openproject')
-  .run(['wpEditField', function(wpEditField) {
-    wpEditField.extendFieldType('select', ['Budget']);
-  }]);
-
-// Register the costs attributes for display
-angular
-  .module('openproject')
-  .run(['wpDisplayField', function(wpDisplayField) {
-    wpDisplayField.extendFieldType('resource', ['Budget']);
-    wpDisplayField.addFieldType(CostsByTypeDisplayField, 'costs', ['costsByType']);
-    wpDisplayField.addFieldType(CurrencyDisplayField, 'currency', ['laborCosts', 'materialCosts', 'overallCosts']);
-  }]);
-
-// main app
-var openprojectCostsApp = angular.module('openproject');
-
-openprojectCostsApp.run(['HookService', function(HookService) {
-  HookService.register('workPackageAttributeEditableType', function(params) {
-    switch (params.type) {
-      case 'Budget':
-        return 'drop-down';
-    }
-    return null;
-  });
-
-  HookService.register('workPackageSingleContextMenu', function(params) {
-    return {
-      key: 'log_costs',
-      icon: 'icon-projects',
-      indexBy: function(actions) {
-          var index = _.findIndex(actions, { key: 'log_time' });
-          return index !== -1 ? index + 1 : actions.length;
-      },
-      resource: 'workPackage',
-      link: 'logCosts'
-    };
-  });
-
-  HookService.register('workPackageTableContextMenu', function(params) {
-    return {
-      key: 'log_costs',
-      icon: 'icon-projects',
-      link: 'logCosts',
-      indexBy: function(actions) {
-        var index = _.findIndex(actions, { link: 'logTime' });
-        return index !== -1 ? index + 1 : actions.length;
-      },
-      text: I18n.t('js.button_log_costs'),
-    };
-  });
-}]);
-
 var requireComponent = require.context('./components/', true, /^((?!\.(test|spec)).)*\.(js|ts)$/);
 requireComponent.keys().forEach(requireComponent);
